+++ libwifi/src/frame/components/mod.rs
@@ -8,4 +8,10 @@ pub use frame_control::FrameControl;
 pub use header::*;
 pub use mac_address::*;
 pub use sequence_control::SequenceControl;
-pub use station_info::StationInfo;
+pub use station_info::{
+    AudioDevices, Cameras, Category, ChannelSwitchAnnouncment, ChannelSwitchMode, Computers,
+    Displays, DockingDevices, GamingDevices, HTInformation, InputDevices, MultimediaDevices,
+    NetworkInfrastructure, PrintersEtAl, RsnAkmSuite, RsnCipherSuite, RsnInformation, StationInfo,
+    Storage, SupportedRate, Telephone, VendorSpecificInfo, WpaAkmSuite, WpaCipherSuite,
+    WpaInformation, WpsInformation, WpsSetupState,
+};