+++ libwifi/src/frame/data/mod.rs
@@ -4,3 +4,15 @@ mod qos_data;
 
 pub use data::*;
 pub use qos_data::*;
+
+use super::components::DataHeader;
+
+pub trait DataFrame {
+    fn header(&self) -> &DataHeader;
+    fn eapol_key(&self) -> &Option<EapolKey>;
+    fn data(&self) -> &Vec<u8>;
+}
+
+pub trait NullDataFrame {
+    fn header(&self) -> &DataHeader;
+}