+++ libwifi/src/frame/management/mod.rs
@@ -1,7 +1,16 @@
+mod action;
 mod association;
+mod authentication;
 mod beacon;
 mod probe;
 
-pub use association::{AssociationRequest, AssociationResponse};
+pub use action::{Action, ActionCategory};
+pub use association::{
+    AssociationRequest, AssociationResponse, Disassociation, ReassociationRequest,
+    ReassociationResponse,
+};
+pub use authentication::{
+    Authentication, Deauthentication, DeauthenticationReason, DEAUTHENTICATION_REASON_MAX,
+};
 pub use beacon::Beacon;
 pub use probe::{ProbeRequest, ProbeResponse};